Data Call refactor - new nexrad call
