+++ src/util/dataCalls/analysis/query-cross-section.ts
@@ -1,4 +1,4 @@
-import { getData } from '../../dataCall'
+import { getData } from '../dataCall'
 
 export const getCrossSectionData = async (site) => {
 	const endpoint = `https://weather.cod.edu/datapoints/analysis/cross-section/get-files.php?parms=${site}`
