+++ src/util/dataCalls/analysis/query-isentropic.ts
@@ -1,4 +1,4 @@
-import { getData } from '../../dataCall'
+import { getData } from '../dataCall'
 
 export const getIsentropicData = async (product) => {
 	const endpoint = `https://weather.cod.edu/datapoints/analysis/rap-mesoanalysis/get-files.php?parms=${product}`
