+++ src/util/dataCalls/analysis/query-soundings.ts
@@ -1,4 +1,4 @@
-import { getData } from '../../dataCall'
+import { getData } from '../dataCall'
 
 export const getSoundingData = async (region, site, product) => {
 	const endpoint = `https://weather.cod.edu/datapoints/analysis/soundings/get-files.php?parms=${region}-${site}-${product}`
