+++ src/util/dataCalls/analysis/query-surface.ts
@@ -1,4 +1,4 @@
-import { getData } from '../../dataCall'
+import { getData } from '../dataCall'
 
 export const getSurfaceData = async (scale, site, product, frames) => {
 	const endpoint = `https://weather.cod.edu/datapoints/analysis/surface/get-files.php?parms=${scale}-${site}-${product}-${frames}`
