+++ src/util/dataCalls/analysis/query-upper-air.ts
@@ -1,4 +1,4 @@
-import { getData } from '../../dataCall'
+import { getData } from '../dataCall'
 
 export const getUpperAirData = async (sector, level, product) => {
 	const endpoint = `https://weather.cod.edu/datapoints/analysis/upper-air/get-files.php?parms=${sector}-${level}-${product}`
